Move taiko `Type` to `Hit` and localise all bind handling
