+++ osu.Game.Rulesets.Taiko/Objects/TaikoStrongableHitObject.cs
@@ -39,12 +39,11 @@ namespace osu.Game.Rulesets.Taiko.Objects
         protected TaikoStrongableHitObject()
         {
             IsStrongBindable.BindValueChanged(_ => updateSamplesFromType());
+            SamplesBindable.BindCollectionChanged((_, __) => updateTypeFromSamples());
         }
 
-        protected override void UpdateTypeFromSamples()
+        private void updateTypeFromSamples()
         {
-            base.UpdateTypeFromSamples();
-
             IsStrong = getStrongSamples().Any();
         }
